requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Jinja2==2.11.2
google-cloud-firestore==1.9.0
requests==2.25.0